divide directory
