+++ redshells/app/word_item_similarity/__init__.py
@@ -1,5 +1,4 @@
 from redshells.app.word_item_similarity.calculate_document_embedding import CalculateDocumentEmbedding
-from redshells.app.word_item_similarity.calculate_similarity_with_matrix_factorization import CalculateSimilarityWithMatrixFactorization
 from redshells.app.word_item_similarity.calculate_word_embedding import CalculateWordEmbedding, CalculateWordEmbeddingWithSCDV
 from redshells.app.word_item_similarity.calculate_word_item_similarity import CalculateWordItemSimilarity
 from redshells.app.word_item_similarity.dimension_reduction_model import TrainDimensionReductionModel, ApplyDimensionReductionModel
