+++ redshells/contrib/model/__init__.py
@@ -1,0 +1,5 @@
+from redshells.contrib.model.factorization_machine import FactorizationMachineGraph, FactorizationMachine
+from redshells.contrib.model.graph_convolutional_matrix_completion import GraphConvolutionalMatrixCompletion
+from redshells.contrib.model.matrix_factorization_model import MatrixFactorizationGraph, MatrixFactorization
+import redshells.model.utils
+import redshells.contrib.model.utils
