+++ redshells/contrib/model/early_stopping.py
