+++ redshells/contrib/model/matrix_factorization_model.py
@@ -5,7 +5,7 @@ import numpy as np
 import sklearn
 
 import redshells
-from redshells.model.early_stopping import EarlyStopping
+from redshells.contrib.model.early_stopping import EarlyStopping
 
 logger = getLogger(__name__)
