+++ redshells/contrib/train/__init__.py
@@ -1,0 +1,3 @@
+from redshells.contrib.train.train_graph_convolutional_matrix_completion import TrainGraphConvolutionalMatrixCompletion
+from redshells.contrib.train.train_matrix_factorization import TrainMatrixFactorization
+import redshells.train.utils
