+++ redshells/model/__init__.py
@@ -1,8 +1,5 @@
-from redshells.model.factorization_machine import FactorizationMachineGraph, FactorizationMachine
 from redshells.model.feature_aggregation_similarity_model import FeatureAggregationSimilarityModel
-from redshells.model.graph_convolutional_matrix_completion import GraphConvolutionalMatrixCompletion
 from redshells.model.lda_model import LdaModel
-from redshells.model.matrix_factorization_model import MatrixFactorizationGraph, MatrixFactorization
 from redshells.model.scdv import SCDV
-import redshells.model.utils
 from redshells.model.tfidf import Tfidf
+import redshells.model.utils
